% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lung}
\alias{lung}
\title{Lung cancer data set}
\format{
An object of class \code{data.frame} with 137 rows and 10 columns.
}
\usage{
lung
}
\description{
Survival in patients with lung cancer presented in Appendix of
Kalbfleisch and Prentice (1980)
}
\details{
\describe{
\item{Therapy}{Type of treatment: standard or test}
\item{Cell}{Cell type}
\item{SurvTime}{Failure or censoring time}
\item{DiagTime}{Months till randomization}
\item{Age}{Age in years}
\item{Prior}{Prior treatment?:0=no,10=yes}
\item{Treatment}{Treatment indicator: 0=standard,1=test}
\item{censor}{Censor indicator: 1=censor,0=event}
}
}
\references{
Kalbfleisch, J. D., and Prentice, R. L. (1980). The Statistical
Analysis of Failure Time Data. New York: John Wiley & Sons.
}
\keyword{datasets}
