% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-MaxLRtest.R
\name{plot.MaxLR}
\alias{plot.MaxLR}
\title{Graphical Display of Weight Functions}
\usage{
\method{plot}{MaxLR}(x, ...)
}
\arguments{
\item{x}{object of \code{MaxLRtest} function}

\item{...}{additional graphical arguments passed to the plot function}
}
\value{
Plots are produced on the current graphics device
}
\description{
Display weight functions used in the function \code{MaxLRtest}
}
\examples{
# See examples in the help file of function MaxLRtest
}
\seealso{
\code{\link{MaxLRtest}}
}
