% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-pwr2n.NPH.R
\name{summary.NPHpwr}
\alias{summary.NPHpwr}
\title{Summary of the \code{pwr2n.NPH} function}
\usage{
\method{summary}{NPHpwr}(object, ...)
}
\arguments{
\item{object}{object of the \code{pwr2n.NPH} function}

\item{...}{additional arguments passed to the summary function}
}
\value{
No return value. Summary results are printed to Console.
}
\description{
Summarize and print the results of \code{pwr2n.NPH} function
}
\seealso{
\code{\link{pwr2n.NPH}}
}
